test_preprocess: finish porting tests to the new mixin
